Add files via upload
This Java code takes the number of subjects and marks obtained in each subject as input from the user. It then calculates the total marks, average percentage, and assigns a grade based on the average percentage. Finally, it displays the total marks, average percentage, and the corresponding grade to the user.

Please note that this code assumes a grading scale where 90-100% corresponds to an 'A', 80-89% corresponds to a 'B', 70-79% corresponds to a 'C', 60-69% corresponds to a 'D', and anything below 60% corresponds to an 'F'. If your grading scale is different, you will need to modify the `calculateGrade` method accordingly.